+++ rusk/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- .tar.xz and .tar.gz compression for state as compile time features [#740]
+
+### Added
 - Prover service implementation [#410]
 
 ### Changed
@@ -67,6 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add linking between Rusk and Protobuff structs
 - Add build system that generates keys for circuits and caches them.
 
+[#740]: https://github.com/dusk-network/rusk/issues/740
+[#410]: https://github.com/dusk-network/rusk/issues/410
 [#401]: https://github.com/dusk-network/rusk/issues/401
 [#369]: https://github.com/dusk-network/rusk/issues/369
 [#327]: https://github.com/dusk-network/rusk/issues/327